+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 # Changelog
 
 ## master
+* Bug: (GUI) Fixed adding and editing records with `datetime` field
 
 ## 4.17.5
 * Feature: (GUI) Basic network log under `/network`
 * Improvement: add new command `modules download`, bring back old syntax to `modules pull`
-
 
 ## 4.17.4
 * Feature: (GUI) Ability to filter logs by string